Use relative measurements in the CSS
All those pixel-exact specifications were a clear tell-tale of
overfitting to one specific dimension. Testament to this is how bad the
page looked on mobile devices (in particular, the need to scroll right
and see that the background is cut off at the window width).

Start fixing this by using measurements relative to the font size.
This is of course only step one. The next steps will follow in separate
commits.

This change was based on the experiences related in

	http://alistapart.com/article/responsive-web-design

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>