Replace download links with ones we never need to change again
There is a nice feature with GitHub release links: you can use the special
"latest" links that will automatically redirect to the latest release.

As releases are made on top of tags (but are not the same), tags without
releases will not be redirect targets, so we do not have to worry about
temporary tags breaking this new linking scheme.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>