bump-version: look only at the latest release
The API call reporting the latest release does not, in fact, return
pre-releases as feared by this maintainer, but it really reports the
latest official release.

Since we are only interested in the latest release for auto-bumping, and
since the full release list is much larger, let's switch to this API
call.  Note: as that API call returns a single release, not an array of
releases, we lose one level of indentation, and the diff is best viewed
with the -w option.

Suggested by Kenn Herman.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>