bump-version: write also `.url` files for direct links
This change lets `bump-version.js` write out \`.url\` files so that e.g.

	https://gitforwindows.org/latest-64-bit-installer.url

will return the direct link to the latest 64-bit Git for Windows
downloader.

Ideally, this would not just be a file returning the URL as static
content, but instead redirect. However, that is currently not possible
with GitHub Pages (which backs Git for Windows' home page).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>