Update the screenshots
The screenshots we had still showed msysGit-based Git for Windows, and
the resolution was really low.

We're in the third decade of the twenty-first century, and it is high
time that we had some higher-quality, and more up to date screenshots.

While at it, ensure that the screenshots all have the very same
dimensions, this will make for a much more pleasant viewing experience.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>