tag-git: use the new Git for Windows snapshots URL
For almost eight years, Git for Windows' snapshots were hosted at
https://wingit.blob.core.windows.net/files/index.html, i.e. on Azure
Blobs.

As of a combination of PRs [*1*], [*2*],
snapshots are not only now built and deployed via GitHub Actions instead
of Azure Pipelines (and ARM64 artifacts are now included, too), they are
also hosted on GitHub [*3*], with the main page being hosted on GitHub
Pages [*4*].

Therefore, the original link now redirects to a new location (which is
also a lot easier to remember): http://gitforwindows.org/git-snapshots.
Let's adjust the link to link there directly.

References:
*1*: https://github.com/git-for-windows/git-for-windows-automation/pull/109
*2*: https://github.com/git-for-windows/gfw-helper-github-app/pull/117
*3*: https://github.com/git-for-windows/git-snapshots/releases/
*4*: https://github.com/git-for-windows/git-snapshots/commits/gh-pages

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>