Allow Open Sans to be used when testing locally
When a web font's URL is specified via `href="//<host>"`, the protocol
is inherited from the loaded page. That's all good and fine if `https`
(or even `http`) is used to load the page, but falls apart when the
protocol is `file`, as it is the case when testing locally.

Therefore, let's just use `https` instead. This should be used nowadays
instead of `http`, anyways.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>