Avoid a semi-clash between the two "32-bit" pages
There is a "32-bit" page that describes the timeline of Git for Windows'
support for i686 (or more precisely, the timeline of _phasing out_ said
support). There is also a page that describes what problems users run
into when they chose to run the 32-bit version of Git for Windows (now a
bit less relevant because Git for Windows won't produce any new 32-bit
installers). Let's avoid confusion by having two so similarly-named
pages, and instead rename the latter to make it more obvious what it is
all about.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>