Stop claiming that v2.48.1 is the latest version
In 4b51fcc (Migrate the home page to Hugo (and to Markdown),
2025-03-10), I had migrated also the header that talks about the current
Git for Windows version and where to find it.

Unfortunately, I forgot to update this partial template to take the
variable data from the site parameters, and instead still hard-coded
the version number and similar information.

So let's fix this ;-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>