replace fancybox with baguettebox and drop jQuery
We mainly used jQuery because fancybox depended on it. But both our
version of fancybox and our version of jquery where severely outdated.
Current versions of fancybox are only available under a commercial
license, so we switch to an open source alternative. Baguettebox was
chosen because it's small, actively maintained, has no dependencies and
offers all the features we used in fancybox.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>