Really avoid any jQuery construct
In https://github.com/git-for-windows/git-for-windows.github.io/pull/76,
I merged a contribution that purported to fix the problem that the
download button no longer pointed to the appropriate installer.

The underlying root cause is that that logic relied on jQuery, but we
recently got rid of it (because our copy was ridiculously outdated).

The problem with that PR is that it had an incomplete solution, still
leaving `$(...)` calls in place.

This here commit fixes that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>